fix(persistence): patch sha1-asm for Apple Silicon compatibility
The sha1-asm crate v0.5.3 has buggy ARM64 assembly that mixes ELF
(:lo12:) and Mach-O (@PAGEOFF) syntax, causing compilation failures
on Apple Silicon with nix-wrapped or recent clang versions.

This adds a local patch that replaces the assembly with a pure Rust
SHA-1 compress function implementation. While slightly slower than
hand-optimized assembly, it provides reliable builds across all
platforms.

Also fixes minor type errors introduced during logging refactor.

+++ apps/persistence/src/workspace/mod.rs
@@ -195,7 +195,7 @@ pub struct Db {
 impl Db {
     pub fn new() -> Self {
         const DB_PATH: &str = ".db";
-        const DB_SIZE: usize = 4 * 1024 * 1024 * 1024; // 4 GiB
+        const DB_SIZE: u64 = 4 * 1024 * 1024 * 1024; // 4 GiB
 
         info!(
             path = DB_PATH,
